+++ tests/test_backends.py
@@ -4,13 +4,13 @@ from unittest import mock
 
 from django.conf import settings
 from django.core.files.images import File
-from django.core.mail import EmailMultiAlternatives, send_mail, EmailMessage
+from django.core.mail import EmailMessage, EmailMultiAlternatives, send_mail
 from django.core.mail.backends.base import BaseEmailBackend
 from django.test import TestCase
 from django.test.utils import override_settings
 
-from ..models import Email, STATUS, PRIORITY
-from ..settings import get_backend
+from post_office.models import PRIORITY, STATUS, Email
+from post_office.settings import get_backend
 
 
 class ErrorRaisingBackend(BaseEmailBackend):
@@ -38,7 +38,7 @@ class BackendTest(TestCase):
     def test_email_backend_setting(self):
         """ """
         old_email_backend = getattr(settings, 'EMAIL_BACKEND', None)
-        old_post_office_backend = getattr(settings, 'POST_OFFICE_BACKEND', None)
+        getattr(settings, 'POST_OFFICE_BACKEND', None)
         if hasattr(settings, 'EMAIL_BACKEND'):
             delattr(settings, 'EMAIL_BACKEND')
         if hasattr(settings, 'POST_OFFICE_BACKEND'):